:fire: Remove test_api_keys.py
