+++ .vscode/settings.json
@@ -2,6 +2,7 @@
   "cSpell.words": [
     "dotenv",
     "lako",
+    "pytest",
     "q",
     "q lako"
   ],
